+++ packages/langium/src/grammar/type-system/ast-collector.ts
@@ -6,7 +6,7 @@
 
 import { Grammar } from '../generated/ast';
 import { LangiumDocuments } from '../../workspace/documents';
-import { AstTypes, InterfaceType, UnionType } from './types-util';
+import { AstTypes, InterfaceType, sortInterfacesTopologically, UnionType } from './types-util';
 import { collectTypeResources } from './type-collector/all-types';
 
 /**
@@ -18,7 +18,7 @@ export function collectAst(documents: LangiumDocuments, grammars: Grammar[]): As
     const { inferred, declared } = collectTypeResources(documents, grammars);
 
     const interfaces: InterfaceType[] = mergeAndRemoveDuplicates({ inferred: inferred.interfaces, declared: declared.interfaces });
-    sortInterfaces(interfaces);
+    sortInterfacesTopologically(interfaces);
 
     const unions: UnionType[] = mergeAndRemoveDuplicates({ inferred: inferred.unions, declared: declared.unions });
     unions.sort((a, b) => a.name.localeCompare(b.name));
@@ -30,36 +30,4 @@ function mergeAndRemoveDuplicates<T extends { name: string }>({ inferred, declar
     return Array.from(inferred.concat(declared)
         .reduce((acc, type) => { acc.set(type.name, type); return acc; }, new Map<string, T>())
         .values());
-}
-
-/**
- * Performs topological sorting on the generated interfaces.
- * @param interfaces The interfaces to sort topologically.
- * @returns A topologically sorted set of interfaces.
- */
-function sortInterfaces(interfaces: InterfaceType[]): InterfaceType[] {
-    type TypeNode = {
-        value: InterfaceType;
-        nodes: TypeNode[];
-    }
-
-    const nodes: TypeNode[] = interfaces
-        .sort((a, b) => a.name.localeCompare(b.name))
-        .map(e => <TypeNode>{ value: e, nodes: [] });
-    for (const node of nodes) {
-        node.nodes = nodes.filter(e => node.value.realSuperTypes.has(e.value.name));
-    }
-    const l: TypeNode[] = [];
-    const s = nodes.filter(e => e.nodes.length === 0);
-    while (s.length > 0) {
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        const n = s.shift()!;
-        if (!l.includes(n)) {
-            l.push(n);
-            nodes
-                .filter(e => e.nodes.includes(n))
-                .forEach(m => s.push(m));
-        }
-    }
-    return l.map(e => e.value);
 }
